v1.882
Pathing was facing the wrong direction at edges and when capping the
path the last non-cap would point out of the room. Checking for when
scouts should switch positions with someone had problems with the
location used, what direction it was checking and other assumptions that
were corrected.  Gather transfer was checking out of bounds when gather
got close to edge of map and so capped checks. Spawner insisted on
having pathLength defined and so gave default value.