1.92
Builders stall if killed before their time, delete unneeded memory on
respawn, clean up memory to keep things sane, pathing ignores flags with
creeps already on it, pathing creates a flag at the start of a path
instead of 1 node in, context holds if the start is blocked it won't be
created and pointed to by the existing path or not.  Scouts use
pathLength and distanceMoved to try to give proper lengths to subsequent
units, removed infinite loop from checking for a new room, scouts forgot
to check if they got a 'scout ahead of me died' and recalculate state in
the room, this was implemented, a lot of scout cpu checks to help
optimization, removed some requestScout checks that were causing
overflow of scouts, added check to look through memory for pre-existing
units since paths being created late or not at all were causing
overflows. Got gatherers to return to get more energy when room is full.