v1.93
Builders will use links if they are near to refill, added a buildRatio
that builds when builders run out of task to gradually repair the lowest
items first (that aren't roads). Added another way for clearing out
scout data to remove flags if room data exists for them, extended
reporting of scouts to 4 hours instead of 1. Made the scout only request
a new unit if the successful storage is confirmed, made storage deny
routes that have the same exit as what was just input, input a way that
in theory will build links next to harvesters that won't be in the way
but are disabled for now, are triggered every 1000 ticks to check if one
is needed near controller or source. Harvesters will use links if they
can and those links when receiving power will try to offload to a less
filled one.