v1.99
Get builder after going to storage to abandon it and go for spawn
instead if out of energy, when a storage exists, only send 1 builder to
controller instead of 2. Convert checking for range to custom distance
function in order to find out what is truely closer to a location.
CreateSpawn creates spawn, observer and power spawn when appropriate
when range of spawn. Convert spawn to allow multiple to exist in a room,
first is labeled master and calls others to support when that one is in
the middle of spawning.