2.15
Major revisions to how scouts track where theyve been so they will pass
through rooms without depending on scout == 0 and if a room is empty of
creeps or not.  Revisions to attackPower to allow either melee or ranged
attackers to build gathers upon near death of the bank and melee creates
healers for themselves. Ranged is the current default.  Moved storage to
activate on room level 4 instead of bundled with links at 5 and stopped
relying and both existing for both to work. Gathers will try to pick up
power and dump it into power spawn and towers and they will try to fill
up power spawn and towers with energy if storage is full enough.
Average cpu cleans itself every 10000 ticks and reports its averages
twice a day. All other reporting moved to twice a day.  Observers will
now scan within a 11x11 grid and store found data within it.