v2.16
Defense has a threat.toString equivalent to allow more informative
reporting. Error reporting for scouting functionality. has a new .tower
function for when towers exist that attacks passed in enemies.
Translated most of storeBank so that it doesn't rely on a unit so the
observer can use it later for found banks.  Units going to storage will
fill up optional structures with power if they aren't full as long as
storage has enough in reserve.  Gatherers handle power now.