v2.418
Added a few goofy unicodes for creeps to say when they're doing stuff.
Re-ordered spawner to try to spawn units first (if no defense needed)
before going into a respawn check state. This state is paused until the
next tick if the current respawn unit is invalid for some reason and we
need to skip to the next unit. This should prevent units from spawning
that aren't needed.